Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -3,7 +3,7 @@
 int main() {
     char ch = 'A';
     scanf("%d",&ch);
-    printf("ASCII Value: %d\n", ch, (int)ch);
+    printf("ASCII Value: %d\n", ch,ch);
 
     return 0;
 }
